+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 9.7.5
+- Fixes an issue, where the PayLater button is not displayed on a product Page when the cart is empty (shopware/shopware#8356)
 - Fixes an issue, where vaulted payments could not be processed in the after order payment process in certain situations
 
 # 9.7.4